lib: include README as doc
Caveats:
Further referenced local markdown files are not included automatically.
While it would be possible to add more `#![doc = ...]` for `docs/*md`,
they would simply be appended to the README, so that links would only work
as #-references. Also, images are not included. Those are known issues
in Rustdoc, and existing workarounds would be complicated to add.
We expect readers to look at the repository's markdown rendering instead.

Signed-off-by: Daniel Maslowski <info@orangecms.org>