me: implement scan() helper
This is for the case that an image is provided that is not as expected.
For upgrade images, some vendors do not declare regions (partitions) as
if they were when residing in flash, probably applied with a certain
logic during the upgrade. This has been encountered with samples from
MSI Cyborg 15 and 17 devices; see the respective support downloads:
https://www.msi.com/Laptop/Cyborg-15-A12VX
https://www.msi.com/Laptop/Cyborg-17-B2RWX

Signed-off-by: Daniel Maslowski <info@orangecms.org>